+++ build.sh
@@ -1,11 +1,13 @@
 set -e
 
 echo adding libraries...
-python build_lib.py
+python lib/build.py
 
-echo linting...
+echo linting code...
 go vet .
 
 echo building binary...
 [ ! -d "./bin" ] && mkdir bin
 go build -o ./bin/fizz.exe run.go main.go
+
+echo "done!"