+++ dev/build_all.sh
@@ -1,6 +1,7 @@
 set -e
 [ ! -d "./bin" ] && mkdir bin
-python build_lib.py
+python lib/build.py
+python lib/autodocs.py
 go vet .
 GOOS=windows GOARCH=amd64 go build -o ./bin/fizz.exe .
 GOOS=linux GOARCH=amd64 go build -o ./bin/fizz .