+++ dev/commit.sh
@@ -4,5 +4,6 @@ go mod tidy
 go mod verify
 go vet .
 gofmt -w -s .
+python lib/autodocs.py
 git add .
 git commit -m "$1"